Workaround for GitHub REST API bug/limitation.
GitHub REST API fails with `500 Internal Server Error`
when trying to use it to create a ref from another repository
where it doesn't have a known relationship to the other repository.

So this workaround does the `git` commands locally in the checked out
repo, using `git fetch` to fetch the base/source repository and checkout
the commit into a new branch and push the new branch to the origin/target repository.